pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragmark"
version = "1.1.0"
description = "Lightweight, configurable context bookmarking layer for RAG pipelines."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Rishabh"},
]

dependencies = [
    "pyyaml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["ragmark*"]
